requirements.txt
expandvars==0.11.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
fastapi==0.109.2
dependency-injector==4.41.0
pydantic~=2.6.1
starlette~=0.36.3
uvicorn==0.27.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
bcrypt==4.0.1
requests==2.31.0
